requirements.txt
asgiref==3.4.1
attrs==21.2.0
bleach==3.1.4
certifi==2021.5.30
chardet==3.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-on-heroku==1.1.2
django-picklefield==2.1.1
django-taggit==1.2.0
entrypoints==0.3
gunicorn==20.1.0
idna==2.10
ipython-genutils==0.2.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-core==4.8.1
lxml==4.5.2
Markdown==3.1.1
MarkupSafe==2.0.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.1.3
pandocfilters==1.5.0
Pillow==7.0.0
psycopg2-binary==2.8.6
PyGithub==1.51
Pygments==2.10.0
PyJWT==2.1.0
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.23.0
six==1.16.0
sqlparse==0.4.2
testpath==0.5.0
traitlets==5.1.0
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.12.1
